refactor: update imports and restructure storage management; add DataDeduplicator and PaginationController classes
