--- apps/extension/lib/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './cookie'
-export * from './storage'
